Change sig of Connection to conform with InformativeConnection
Sqlite3 returns an int64 in sqlite3_last_rowid (which gets the last
inserted row's primary key). Unfortunately, the db API requires us to
return an integer to conform, so just clip off the trailing bits.

It may be prudent to change the API and require other implementations to
add additonal bits on, but bleh.